[#23] Rename args files to scc_args
